apps/ai-service/main.py
from fastapi import FastAPI, HTTPException, UploadFile, File, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field
from typing import Optional, List, Dict, Any
from datetime import datetime, date
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="CombatID AI Service",
    description="AI-powered document intelligence for combat sports",
    version="0.1.0",
)

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=os.getenv("CORS_ORIGINS", "http://localhost:3000,http://localhost:3001").split(","),
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


# Models
class DocumentType(str):
    BLOOD_TEST = "blood_test"
    PHYSICAL = "physical"
    EYE_EXAM = "eye_exam"
    MRI = "mri"
    EKG = "ekg"
    DRUG_TEST = "drug_test"
    LICENSE = "license"
    INSURANCE = "insurance"
    OTHER = "other"


class ExtractedMetadata(BaseModel):
    document_type: str
    issue_date: Optional[date] = None
    expiration_date: Optional[date] = None
    provider: Optional[str] = None
    patient_name: Optional[str] = None
    results: Optional[Dict[str, Any]] = None
    raw_text: Optional[str] = None


class DocumentAnalysisResult(BaseModel):
    document_id: str
    classification: str
    confidence: float = Field(ge=0, le=1)
    extracted_data: ExtractedMetadata
    flags: List[str] = []
    requires_review: bool = False


class ProcessDocumentRequest(BaseModel):
    document_id: str
    file_url: str
    fighter_id: str
    expected_type: Optional[str] = None


class HealthCheckResponse(BaseModel):
    status: str
    version: str
    services: Dict[str, bool]


# Routes
@app.get("/", response_model=HealthCheckResponse)
async def health_check():
    """Health check endpoint"""
    return HealthCheckResponse(
        status="healthy",
        version="0.1.0",
        services={
            "ocr": True,
            "classification": True,
            "extraction": True,
        }
    )


@app.post("/api/v1/documents/analyze", response_model=DocumentAnalysisResult)
async def analyze_document(request: ProcessDocumentRequest):
    """
    Analyze a document using AI:
    1. OCR to extract text
    2. Classify document type
    3. Extract structured metadata
    4. Flag anomalies
    """
    # TODO: Implement actual AWS Textract / OpenAI integration
    # For now, return mock data

    mock_result = DocumentAnalysisResult(
        document_id=request.document_id,
        classification=request.expected_type or "blood_test",
        confidence=0.95,
        extracted_data=ExtractedMetadata(
            document_type=request.expected_type or "blood_test",
            issue_date=date.today(),
            expiration_date=date(2026, 6, 1),
            provider="Quest Diagnostics",
            patient_name="John Doe",
            results={
                "hiv": "negative",
                "hepatitis_b": "negative",
                "hepatitis_c": "negative",
            },
            raw_text="[OCR extracted text would appear here]"
        ),
        flags=[],
        requires_review=False,
    )

    return mock_result


@app.post("/api/v1/documents/classify")
async def classify_document(file: UploadFile = File(...)):
    """
    Classify an uploaded document without full processing.
    Quick classification for UI feedback.
    """
    # TODO: Implement lightweight classification
    content_type = file.content_type
    filename = file.filename or ""

    # Simple heuristic classification based on filename
    classification = "other"
    confidence = 0.5

    filename_lower = filename.lower()
    if "blood" in filename_lower or "lab" in filename_lower:
        classification = "blood_test"
        confidence = 0.8
    elif "physical" in filename_lower or "exam" in filename_lower:
        classification = "physical"
        confidence = 0.8
    elif "eye" in filename_lower or "vision" in filename_lower:
        classification = "eye_exam"
        confidence = 0.8
    elif "mri" in filename_lower or "scan" in filename_lower:
        classification = "mri"
        confidence = 0.8
    elif "ekg" in filename_lower or "ecg" in filename_lower:
        classification = "ekg"
        confidence = 0.8
    elif "drug" in filename_lower:
        classification = "drug_test"
        confidence = 0.8
    elif "license" in filename_lower or "id" in filename_lower:
        classification = "license"
        confidence = 0.8

    return {
        "classification": classification,
        "confidence": confidence,
        "filename": filename,
        "content_type": content_type,
    }


@app.post("/api/v1/documents/ocr")
async def extract_text(file: UploadFile = File(...)):
    """
    Extract text from a document using OCR.
    Supports PDF and images.
    """
    # TODO: Implement AWS Textract integration
    return {
        "text": "[OCR text extraction would appear here]",
        "pages": 1,
        "confidence": 0.9,
    }


@app.post("/api/v1/documents/validate")
async def validate_document(request: ProcessDocumentRequest):
    """
    Validate document against fighter profile.
    Check for:
    - Name matching
    - Date validity
    - Document tampering signs
    """
    # TODO: Implement validation logic
    return {
        "document_id": request.document_id,
        "fighter_id": request.fighter_id,
        "validation_passed": True,
        "checks": {
            "name_match": {"passed": True, "confidence": 0.95},
            "date_valid": {"passed": True},
            "tampering_detected": {"passed": True, "confidence": 0.98},
        },
        "flags": [],
    }


@app.post("/api/v1/eligibility/calculate")
async def calculate_eligibility(
    fighter_id: str,
    commission_id: str,
    discipline: str,
    documents: List[Dict[str, Any]],
):
    """
    Calculate fighter eligibility based on documents and ruleset.
    Returns detailed eligibility breakdown.
    """
    # TODO: Implement eligibility calculation
    return {
        "fighter_id": fighter_id,
        "commission_id": commission_id,
        "discipline": discipline,
        "status": "eligible",
        "requirements": [],
        "expiring_soon": [],
        "calculated_at": datetime.now().isoformat(),
    }


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
