apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.10.0
pydantic-settings==2.6.0
boto3==1.35.0
openai==1.55.0
httpx==0.28.0
python-jose[cryptography]==3.3.0
Pillow==11.0.0
pdf2image==1.17.0
pytesseract==0.3.13
python-dotenv==1.0.1
